+++ src/API/Grand.Api/Commands/Handlers/Catalog/DeleteProductAttributeCommandHandler.cs
@@ -4,32 +4,28 @@ using Grand.Business.Core.Interfaces.Common.Localization;
 using Grand.Infrastructure;
 using MediatR;
 
-namespace Grand.Api.Commands.Handlers.Catalog
+namespace Grand.Api.Commands.Handlers.Catalog;
+
+public class DeleteProductAttributeCommandHandler : IRequestHandler<DeleteProductAttributeCommand, bool>
 {
-    public class DeleteProductAttributeCommandHandler : IRequestHandler<DeleteProductAttributeCommand, bool>
+    private readonly IProductAttributeService _productAttributeService;
+    private readonly ITranslationService _translationService;
+    private readonly IWorkContext _workContext;
+
+    public DeleteProductAttributeCommandHandler(
+        IProductAttributeService productAttributeService,
+        ITranslationService translationService,
+        IWorkContext workContext)
     {
-        private readonly IProductAttributeService _productAttributeService;
-        private readonly ITranslationService _translationService;
-        private readonly IWorkContext _workContext;
+        _productAttributeService = productAttributeService;
+        _translationService = translationService;
+        _workContext = workContext;
+    }
 
-        public DeleteProductAttributeCommandHandler(
-            IProductAttributeService productAttributeService,
-            ITranslationService translationService,
-            IWorkContext workContext)
-        {
-            _productAttributeService = productAttributeService;
-            _translationService = translationService;
-            _workContext = workContext;
-        }
-
-        public async Task<bool> Handle(DeleteProductAttributeCommand request, CancellationToken cancellationToken)
-        {
-            var productAttribute = await _productAttributeService.GetProductAttributeById(request.Model.Id);
-            if (productAttribute != null)
-            {
-                await _productAttributeService.DeleteProductAttribute(productAttribute);
-            }
-            return true;
-        }
+    public async Task<bool> Handle(DeleteProductAttributeCommand request, CancellationToken cancellationToken)
+    {
+        var productAttribute = await _productAttributeService.GetProductAttributeById(request.Model.Id);
+        if (productAttribute != null) await _productAttributeService.DeleteProductAttribute(productAttribute);
+        return true;
     }
 }
