+++ src/API/Grand.Api/Commands/Handlers/Catalog/DeleteProductSpecificationCommandHandler.cs
@@ -2,30 +2,26 @@
 using Grand.Business.Core.Interfaces.Catalog.Products;
 using MediatR;
 
-namespace Grand.Api.Commands.Handlers.Catalog
+namespace Grand.Api.Commands.Handlers.Catalog;
+
+public class DeleteProductSpecificationCommandHandler : IRequestHandler<DeleteProductSpecificationCommand, bool>
 {
-    public class DeleteProductSpecificationCommandHandler : IRequestHandler<DeleteProductSpecificationCommand, bool>
+    private readonly IProductService _productService;
+    private readonly ISpecificationAttributeService _specificationAttributeService;
+
+    public DeleteProductSpecificationCommandHandler(ISpecificationAttributeService specificationAttributeService,
+        IProductService productService)
     {
-        private readonly ISpecificationAttributeService _specificationAttributeService;
-        private readonly IProductService _productService;
+        _specificationAttributeService = specificationAttributeService;
+        _productService = productService;
+    }
 
-        public DeleteProductSpecificationCommandHandler(ISpecificationAttributeService specificationAttributeService,
-            IProductService productService)
-        {
-            _specificationAttributeService = specificationAttributeService;
-            _productService = productService;
-        }
+    public async Task<bool> Handle(DeleteProductSpecificationCommand request, CancellationToken cancellationToken)
+    {
+        var product = await _productService.GetProductById(request.Product.Id, true);
+        var psa = product.ProductSpecificationAttributes.FirstOrDefault(x => x.Id == request.Id);
+        if (psa != null) await _specificationAttributeService.DeleteProductSpecificationAttribute(psa, product.Id);
 
-        public async Task<bool> Handle(DeleteProductSpecificationCommand request, CancellationToken cancellationToken)
-        {
-            var product = await _productService.GetProductById(request.Product.Id, true);
-            var psa = product.ProductSpecificationAttributes.FirstOrDefault(x => x.Id == request.Id);
-            if (psa != null)
-            {
-                await _specificationAttributeService.DeleteProductSpecificationAttribute(psa, product.Id);
-            }
-
-            return true;
-        }
+        return true;
     }
 }
