+++ src/API/Grand.Api/Commands/Handlers/Common/DeletePictureCommandHandler.cs
@@ -2,27 +2,22 @@
 using Grand.Business.Core.Interfaces.Storage;
 using MediatR;
 
-namespace Grand.Api.Commands.Handlers.Common
+namespace Grand.Api.Commands.Handlers.Common;
+
+public class DeletePictureCommandHandler : IRequestHandler<DeletePictureCommand, bool>
 {
-    public class DeletePictureCommandHandler : IRequestHandler<DeletePictureCommand, bool>
+    private readonly IPictureService _pictureService;
+
+    public DeletePictureCommandHandler(IPictureService pictureService)
     {
-        private readonly IPictureService _pictureService;
+        _pictureService = pictureService;
+    }
 
-        public DeletePictureCommandHandler(IPictureService pictureService)
-        {
-            _pictureService = pictureService;
-        }
+    public async Task<bool> Handle(DeletePictureCommand request, CancellationToken cancellationToken)
+    {
+        var picture = await _pictureService.GetPictureById(request.PictureDto.Id);
+        if (picture != null) await _pictureService.DeletePicture(picture);
 
-        public async Task<bool> Handle(DeletePictureCommand request, CancellationToken cancellationToken)
-        {
-            var picture = await _pictureService.GetPictureById(request.PictureDto.Id);
-            if (picture != null)
-            {
-                await _pictureService.DeletePicture(picture);
-            }
-
-            return true;
-
-        }
+        return true;
     }
 }
